Update build script

+++ build.sh
@@ -16,14 +16,74 @@ get_object_info_from_filename() {
 
 }
 
-call_sqlcmd_v2() {
+call_sqlcmd() {
 
-    result=$(sqlcmd -u -W -w 999 -h-1 -s"," -i input.sql | tr -d '\r')
-    echo "$result" > file1
+    if [ "$1" = "column" ]; then
+        result=$(sqlcmd -u -W -w 999 -h-1 -s"," -i input.sql | tr -d '\r')
+        echo "$result" > file_from_db
+    elif [ "$1" = "index" ]; then
+        result=$(sqlcmd -u -W -w 999 -h-1 -s"," -i input.sql | tr -d '\r')
+        echo "$result"  > current_index_in_db
+        echo "$result"  | grep -oP '(\(.+\))WITH (\(.+\))' | sed -e 's/WITH/,/g' | sed -e 's/(//g' | sed -e 's/)//g' | tr ',' '\n' > file_from_db
+    fi
+
 
 }
 
-get_table_info_from_db_v3() {
+get_table_index_info_from_db() {
+
+    cat > input.sql << EOM
+:setvar SQLCMDERRORLEVEL 1
+USE DEV
+GO
+SET NOCOUNT ON
+GO
+SELECT
+   CONCAT( 
+   'CONSTRAINT ',
+   QUOTENAME(i.name),
+   ' ',
+   'PRIMARY KEY',
+   ' ',
+   i.type_desc COLLATE DATABASE_DEFAULT,
+   ' (  ',
+   STRING_AGG(QUOTENAME(c.name) + ' ASC',','),
+   ' )WITH (',
+   'PAD_INDEX = ',
+   case when i.is_padded = 1 then 'ON, ' else 'OFF, ' end,
+   'STATISTICS_NORECOMPUTE = ',
+   case is_padded when 1 then 'ON, ' else 'OFF, ' end,
+   'IGNORE_DUP_KEY = ',
+   case when ignore_dup_key = 1 then 'ON, ' else 'OFF, ' end,
+   'ALLOW_ROW_LOCKS = ',
+   case when allow_row_locks = 1 then 'ON, ' else 'OFF, ' end,
+   'ALLOW_PAGE_LOCKS = ',
+   case when allow_page_locks = 1 then 'ON, ' else 'OFF, ' end,
+   'OPTIMIZE_FOR_SEQUENTIAL_KEY = ',
+   case when optimize_for_sequential_key = 1 then 'ON, ' else 'OFF' end,
+   ') ON [PRIMARY]'
+   )
+FROM
+    sys.tables t
+INNER JOIN 
+    sys.indexes i ON t.object_id = i.object_id
+INNER JOIN 
+    sys.index_columns ic ON i.index_id = ic.index_id AND i.object_id = ic.object_id
+INNER JOIN 
+    sys.columns c ON ic.column_id = c.column_id AND ic.object_id = c.object_id
+WHERE
+    i.index_id = 1  -- clustered index
+	and t.name = '$2'
+group by i.name, t.schema_id, t.name, i.type_desc, i.is_padded, i.ignore_dup_key, 
+i.allow_row_locks, i.allow_page_locks, i.optimize_for_sequential_key
+GO
+EOM
+
+    call_sqlcmd "index"
+
+}
+
+get_table_column_info_from_db() {
 
     cat > input.sql << EOM
 :setvar SQLCMDERRORLEVEL 1
@@ -53,22 +113,36 @@ where TABLE_SCHEMA = '$1' AND TABLE_NAME = '$2'
 GO
 EOM
 
-    call_sqlcmd_v2
+    call_sqlcmd "column"
 
 }
 
-get_table_info_from_file_v2() {
+get_table_index_info_from_file() {
+
+    result=$(sed -n '/CONSTRAINT/{/NONCLUSTERED/q; p; :loop n; p; /WITH/q; b loop}' "$1" | tr '\n' ' ' )
+    echo "$result" | grep -oP -m 1 '(\[[a-zA-Z_]+\])' > index_name
+    echo "$result" | grep -oP '(\(.+\))WITH (\(.+\))' | sed -e 's/WITH/,/g' | sed -e 's/(//g' | sed -e 's/)//g' | tr ',' '\n' > file_from_repo
+
+}
+
+get_table_column_info_from_file() {
 
     result=$(sed -e '/CREATE/,/GO/!d' "$1" | grep -e ',' | grep -v 'ASC,' | grep -v ')WITH' | sed -e 's/^\s*//')
-    echo "$result" > file2
+    echo "$result" > file_from_repo
 
 }
 
-find_difference_v2() {
+find_difference() {
 
-    result=$(diff file1 file2)
-    echo "$result"
+    if [ -z ${1} ]; then
+        result=$(diff file_from_db file_from_repo)
+        echo "$result"
+    elif [ "$1" = "ignore_space" ]; then
+        result=$(diff -w file_from_db file_from_repo)
+        echo "$result"
+    fi
 
+    rm file_from_db file_from_repo
 }
 
 generate_alter_statement() {
@@ -87,29 +161,57 @@ generate_alter_statement() {
         echo "$repo_line"
     fi
 
-    ALTER_STATEMENT=$(echo "ALTER TABLE "$1.$2)
-    echo "$ALTER_STATEMENT" >> database.sql
+    if [ "$4" = "column" ]; then
+        ALTER_STATEMENT=$(echo "ALTER TABLE [$1].[$2]")
+        echo "$ALTER_STATEMENT" >> database.sql 2>&1
 
-    if [ "$diff_action" == "a" ]; then
-        MODIFY_COLUMN=$(echo "ADD "$object_change)
-        echo "$MODIFY_COLUMN" >> database.sql
-    elif [ "$diff_action" == "d" ]; then 
-        IFS=' ' read -ra OBJECT <<< "$object_change"
-        DROP_COLUMN=$(echo "${OBJECT[0]}")
-        MODIFY_COLUMN=$(echo "DROP COLUMN "$DROP_COLUMN)
-        echo "$MODIFY_COLUMN" >> database.sql
-    elif [ "$diff_action" == "c" ]; then
-        # todo 
-        echo "NOP"
+        if [ "$diff_action" == "a" ]; then
+            MODIFY_COLUMN=$(echo "ADD "$object_change)
+            echo "$MODIFY_COLUMN" >> database.sql 2>&1
+        elif [ "$diff_action" == "d" ]; then 
+            IFS=' ' read -ra OBJECT <<< "$object_change"
+            DROP_COLUMN=$(echo "${OBJECT[0]}")
+            MODIFY_COLUMN=$(echo "DROP COLUMN "$DROP_COLUMN)
+            echo "$MODIFY_COLUMN" >> database.sql 2>&1
+        elif [ "$diff_action" == "c" ]; then
+            # todo 
+            echo "NOP"
+        fi
+    elif [ "$4" = "index" ]; then
+        ALTER_STATEMENT=$(echo "ALTER TABLE [$1].[$2] DROP CONSTRAINT $5 WITH ( ONLINE = OFF )")
+        echo "$ALTER_STATEMENT" >> database.sql 2>&1
+
+        if [ "$diff_action" == "a" ]; then
+            # todo 
+            echo "NOP"
+        elif [ "$diff_action" == "d" ]; then 
+            CURRENT_INDEX=$(cat current_index_in_db)
+            if [ $debug == 1 ]; then
+                echo "current index in db: $CURRENT_INDEX"
+            fi
+            index_to_be_removed=$(echo "$object_change" | awk '{$1=$1};1' | sed 's/[][\/*.]/\\&/g; s%.*%,&%')
+            if [ $debug == 1 ]; then
+                echo "index_to_be_removed: $index_to_be_removed"
+            fi
+            MODIFY_INDEX=$(sed -e "s|${index_to_be_removed}||g" current_index_in_db )
+            if [ $debug == 1 ]; then
+                echo "modified index: $MODIFY_INDEX"
+            fi
+            echo "ALTER TABLE [$1].[$2] ADD $MODIFY_INDEX" >> database.sql 2>&1
+        elif [ "$diff_action" == "c" ]; then
+            # todo 
+            echo "NOP"
+        fi
     fi
         
-    echo "GO" >> database.sql
+    echo "GO" >> database.sql 2>&1
 
 }
 
 
 # mode can be full or delta
 # debug can be 1 or 0
+# example: ./build.sh -d 1 -m delta
 while getopts ":d:m:" opt; do
   case $opt in
     d) debug="$OPTARG"
@@ -131,6 +233,7 @@ done
 printf "Argument debug is %s\n" "$debug"
 printf "Argument mode is %s\n" "$mode"
 
+rm -rf build
 mkdir build
 cp dbo.*.sql build
 cd build
@@ -148,33 +251,63 @@ do
         if [ "$object_type" == "Table" ]; then
 
             if [ $debug == 1 ]; then
-                echo "$object_name"
+                echo "---------------------------------------------------------"
+                echo "table: $object_name"
+            fi
+
+            # get table column info from db
+            get_table_column_info_from_db "$object_schema" "$object_name"
+
+            # then we get table column info from repo
+            get_table_column_info_from_file "$file"
+
+            # find the column difference
+            diff_result=$(find_difference)
+
+            if [ $debug == 1 ]; then
+                echo "---------------------------------------------------------"
+                echo "$diff_result"
+            fi
+
+            # generate alter column statement
+            if [ -z "$diff_result" ]; then
+                echo "No column difference"
+            else
+                diffs=$(echo "$diff_result" | paste -sd "," - | sed -e 's/<//g' | sed -e 's/>//g')
+                echo "$diffs"
+                generate_alter_statement $object_schema $object_name "$diffs" "column"
+            fi
+        
+            # get table index info from db
+            get_table_index_info_from_db "$object_schema" "$object_name"
+
+            # then we get table index info from repo
+            get_table_index_info_from_file "$file"
+
+            # grab the index name
+            index_name=$(cat index_name)
+            if [ $debug == 1 ]; then
+                echo "index: $index_name"
                 echo "---------------------------------------------------------"
             fi
 
-            # get table info from db
-            get_table_info_from_db_v3 "$object_schema" "$object_name"
-
-            # then we get table info from repo
-            get_table_info_from_file_v2 "$file"
-
-            # find the difference
-            diff_result=$(find_difference_v2 $db_result $file_result)
+            # find the index difference
+            diff_result=$(find_difference "ignore_space")
 
             if [ $debug == 1 ]; then
                 echo "$diff_result"
                 echo "---------------------------------------------------------"
             fi
 
-            # generate alter statement
+            # generate alter index statement
             if [ -z "$diff_result" ]; then
-                echo "No difference"
+                echo "No index difference"
             else
                 diffs=$(echo "$diff_result" | paste -sd "," - | sed -e 's/<//g' | sed -e 's/>//g')
                 echo "$diffs"
-                generate_alter_statement $object_schema $object_name "$diffs"
+                generate_alter_statement $object_schema "$object_name" "$diffs" "index" $index_name
             fi
-        
+
         # then we get view info from db
             
             # find the difference
@@ -207,6 +340,6 @@ done
 if [ "$mode" == "full" ]; then
     for tmp in table.tmp view.tmp sp.tmp
     do
-        cat $tmp >> database.sql
+        cat $tmp >> database.sql 2>&1
     done
 fi
